TSAPPS-224 Fix mapping for success report
Columns from mapping with spaces or *-symbols in source data are concidered as valid and similar columns
In report this column names was not proceeded as similar - so report had 2 banch of same columns with empty data in relative cells

Also was separated and refactored logic of header building. And Success Report data builder functions were separated to own file.
Added acceptence test coverage for Transformation logic (Success report building)