TSAPPS-310 Add UoM mapping
TS expects to get UOMs in UBL format.
So if sellers have it in a different format that they need to provide mapping on all used UOMs in their feeds.
We need to provide the possibility to apply this mapping and use it on the transformation and use actual values in failures report